checkpoint: renames "/composables/" module to "/ProgressiveRef/"
