+++ src/library/vue/exports.toolbox.ts
@@ -4,8 +4,8 @@ export type * from './component';
 
 export * from './Ref';
 
+export * from './ProgressiveRef';
+
 export * from './Reactive';
 
 export * from './watchers';
-
-export * from './composables';
